+++ CHANGELOG.md
@@ -22,7 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added combined_equipment energy category report hyperlink with equipment uuid to combined-equipment batch report
 - added equipment energy category report hyperlink with equipment uuid to equipment batch report
 - added latitude and longitude to Space in database, myems-api and myems-admin
-- 
+- added run immediately action to Advacned Report in database, myems-api and myems-admin
 ### Changed
 - changed user avator from emoji to image
 ### Fixed